git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/releases/moos-ivp-15.5@7227 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ MOOS_V10.0.3_May1215/MOOSCore/Core/libMOOS/Utils/include/MOOS/libMOOS/Utils/IPV4Address.h
@@ -60,13 +60,13 @@ public:
 	std::string host() const;
 	void set_host(const std::string & host);
 
-    uint16_t port() const;
-    void set_port(uint16_t port);
+    uint32_t port() const;
+    void set_port(uint32_t port);
 
 
 protected:
 	std::string host_;
-    uint16_t port_;
+    uint32_t port_;
 
 };
